Remove JQuery and Coffee script gems
Both these gems are installed as part of the default Rails installation.

As a rule of thumb we avoid complex JavaScript and so It is unlikely we
will need to utilise either, we can remove them and the burden of
keeping them up-to-date.

Should the need for either arise, they can easily be re-introduced.