Ingest planned disbursements
Like transactions, a large number of legacy planned disbursements have
no `type` attribute for providing organisation and receiving
organisation.

As the `type` attribute is required in the application we have to
provide a value. It would be very difficult to calculate the type from
the legacy activity so we will set values to:

- 10 (Government) for providing organisations
- 0 for receiving organisations

for those that do not have the attribute.

The providing organisation type is correct until we begin to ingest
non-government organisations data (which will come in time).

The receiving organisation type is not a recognised IATI organisation
type:

http://reference.iatistandard.org/203/codelists/OrganisationType/