Exporting unknown receiving organisations
Now that planned disbursements can be ingested and their receiving
organisation type can be set to "0" which is not a recognised IATI
organisation type, we should only export the attributes for receiving
organisation if we know they will be valid i.e. that @type is not "0",
we still export the receiving organisation name, so the xml is valid.