Switch our version of the form builder gem to the latest
- During a separate piece of work we have been trying to take advantage
of a new feature of this gem.
- Since we were using a dxw fork, we were unable to do this, so had to
either switch away from it or update our fork. Given that it looked like
the gem was now providing the feature that caused us to originally fork,
we think that it is safe to switch back.
- On updating the gem we found that there were able to remove the monkey
patching for Internationalisation and replace it with gem configuration
https://govuk-form-builder.netlify.app/building-blocks/localisation/#customising-locale-structure
- We found that we needed to continue to provide an override so that we
could continue to use a single set of local values rather than have it
duplicated. This is consist with the initial approach our application
was using.
- Whilst by default the gem looks for locale values under helpers.legend
we were already using helpers.fieldset, so we made another configuration
change to allow us to continue to use helpers.fieldset
- We have changed html_safe? to html_safe without a question mark after
a test failure made us aware of the fact that this hint text was never
being used. Since the update to the gem we found that we do need to have
this line for the hint text appear

Co-authored-by: Tom Hipkin <tomhipkin@gmail.com>