Activities know their providing organisation
When adding a planned disbursement or transaction the providing
organisation can be derived from the type of activity and the delivery
partner who added the activity.

This will allow the service to either not ask the user to provide this
information or pre-fill the information in forms so it can still be
changed when needed.