Pre-fill planned disbursement providing org
User research showed that users did not understand why they have to
enter details that the service should already 'know'.

Now that an Activity can derive it's own providing organisation we can
pre-fill the planned disbursement form with the details of the correct
organisation, aligning with users expectations.

If there is a case we do not yet know about, the user can change the
providing organisation details and save the planned disbursement.