Toggle automatic data migrations off
For restoring production data to other live environments, old data migrations were being run on data that was in a newer state. This raised errors where column names no longer exists and prevented the application from being booted.

Since the container then fails to start during a deployment gpaas rolls back the change and we are left with a crashing application.

We cannot delete existing or create new databases as we do not have a user with the appropriate permissions. This means we have to delete the entire instance and recreate it. This process will give us a new database name which then merge the backfup sql file ontop of.

This change provides a toggle so that we can disable automatic data migrations when setting up a fresh database from a copy. Under normal circumstances data migrations should continue running on production so the default is "true".