(fix) Tidy up old Activities which may not have a reporting_organisation
While working on the pentest environment we discovered some Activities with a nil
reporting_organisation. These records are old and we are no longer able to create
activities without a reporting organisation.

This data migration ensures there are no old records without a reporting organisation.

It follows the same rules as those in the Create[Level]Activity services.

It is irreversible because Activities are not valid without a reporting organisation.