Strip whitespace from Activity identifier value
We have seen an instance where leading whitespace in the IATI
Identifier value caused the XML to fail the IATI validator.

The IATI spec states 'The iati-identifier should have no leading or
trailing whitespace.' therefore we should strip the whitespace before we
save the value.

http://reference.iatistandard.org/203/activity-standard/overview/iati-identifier/

The strip_attributes gem was chosen:

https://github.com/rmm5t/strip_attributes

This was after much discussion around the various approaches could
take. We felt the gem was simple enough and gave us the functionality we
wanted without a large overhead. Under the hood, it uses the before_save
callback, which was one of the preferred approaches, but the gem has a
nice interface to allow targeting the attributes to act upon.

Configure RSpec to use the strip_attribute matchers so we can use them
in our specs.