Improve "create activity" content
Improve the labelling for the Create Activity form steps.

This ticket has involved changing many of the strings used in the Activity
create form. Previously the pate title strings were being used as labels for
the fields/fieldsets in the create activity form, but this is no longer
feasible as the content for page titles & fieldset labels need to be different.