Document the latest way to reset an environments database to production
Since the original documentation the user we were using through conduit no longer has permissions to delete or create databases. I'm not sure if this was an anomaly the first time or something has changed within gpaas since this was written.

As I'm unable to delete and recreate by hand I have been using a more nuclear option of destroying and readding the instance.

The terraform apply command now includes a flag that turns off the data migrate task and instead loads the schema, trusting that the database has had all required migrations run.