Move and update Pundit translation file
Change the file name to better reflect the use to help non-developer
folks who may not know what 'Pundit' is.

Add some documentation to show developers how they might utilise the
file and retain the link to Pundit for those folks.

Update the call to the t() method to incorporate the change.

This commit causes the test suite to fail because the I18n-tasks gem
does not allow comments in the locale files.

After discussion with the developer team, it is felt that the
'normalization' is not overly helpful and so can be removed to allow for
commented and different structured locale files.

The next commit will perform make the change and restore the test-suite
to green as it is felt important to single that change out in the
history.