Locales no longer have to be normalized
The I18n-tasks gem default configuration is to have a failing test if
the locale files are not normalized.

For this application that meant 'sorted alphabetically by key'. Whilst
this sorting has benefits, it also has drawbacks, making locales less
readable for some folk and removing comments.

In the previous commit a comment was added that add some value to the
file but could not be retained with the normalization was still
enforced.

The I18n-tasks unused and missing key features are still in place and
are helpful to the team to manage the locales for the application.

After this change it is no longer required to run the `i18n-tasks
normalize` task although a developer can still run it and mange the
change themselves.