Move page_title translations to files
The page_title is generally used for the browser document title (shown
in the top of the browser window or tab) and is an important
accessibility feature.

Store each page_title in a relevant file, which meant creating some new
files.

Later work will rename these to make the difference between page title
content (on the page) and browser document titles.