Booting the app with missing env variables will give a meaningful error
I noticed that as a result of change to session_store.rb which rqeuired REDIS_URL to be present. This change was order to make sure our session storage was consistent across environments and to fail loudly if it wasn't, removing the chance of obscure errors that were inconsistent to debug. Eg, cookie overflows.

After rolling this change out we found out that existing members of the dev team and new joiners got tripped up by this new requirement.

This change moves to make `rails s` produce a meaninful error to developers so that they can quickly establish that they are missing environment variables.

Since we are using Dotenv to load our environment variables I've used their recommended approach [1] and added what I believe are the required keys for a successful boot and sign in. Skylight, rollbar and google analytics are not required.

There is a trade off that this file will need to be maintained with new environment variables and that if we forget we will run into the same issue. The app is 6 months old now so I don't expect an avalanche of new dependencies that would make this less valuable to include now.

[1] https://github.com/bkeepers/dotenv#required-keys