Successfully ingest AMS Newton activities which do not have an actual_start_date
When attempting to injest AMS Newton fund activities we discovered that 17 of
the activities had activity-dates type 1 & 3 (planned start & planned end),
but not type 2 (actual start). Our ingestion service expects type 2 dates to be
present, so we were failing to ingest these activities.

We now allow all 4 activity-date types to be empty in the IATI XML we ingest,
but the Activity created from the ingest will not be valid on save if the
actual_start_date or planned_start_date are nil (either may be present).