Update Activity tables
With the refactored locales, the tables that display the various
activity levels had to be updated.

As with all the refactoring of the locales, being explicit is favoured
over not repeating keys or values, we hope this will make maintaining
the translation files simpler.

This commit makes the test suite return to green.