A programme's implementing organisation is the same as its extending organisation
When a programme has its extending organisation set, the same organisation is set
as the implementing organisation.

Previously we were not requiring an implementing organisation for programmes, but
if we are going to publish programmes to IATI we will need them to have an
implementing organisation set. This will always be the same as their extending
organisation.