Activity parent relationship now uses the field name `parent_id`
Previously this was called `activity_id` which was a problem for at least 2 reasons that drove this change:

1. `activity.activity_id` is not as clear as it could be that it wants the parent activity record, it could be any activity id, or perhaps even (for some reason) another identifier for this activity
2. When looking at using the a gem called Ancestry to add a tree structure to help us perform larger queries to connect related nodes together, there seems to be a convention [1] around using parent_id. This field name is required for the gems to work.

[1] https://github.com/stefankroes/ancestry#migrating-from-plugin-that-uses-parent_id-column

I could have added parent_id and kept activity_id to keep the change smaller with:

```
  belongs_to :activity, optional: true, foreign_key: "parent_id"
  belongs_to :parent, optional: true, class_name: :Activity, foreign_key: "parent_id"
```

Though given it was possible to switch entirely to parent, I felt it best to do so and remove an additonal concept from the codebase.

The migration requires save with `validate: false` since activities may
be in an 'incomplete' state, where it has been partially completed. The
thinking is that given we are making a single and focused change to a
hidden field name, we should be safe to persist everyone of these
updates regardless of it's state.