Ingest negative budgets as a penny
The legacy IATI data can sometimes contain negative values, which is not
something we want to ingest, although equally we don't want to ignore
these, so we're setting them to a penny (0.01) for the time being.

Setting it to this value also means we don't need to modify the existing
validations, which would also affect users entering data through the
frontend.