Refactor fetching of activity parents by using a acts_as_tree
This greatly simplifies the parent_activities method. It is no longer dependent on having to interograte the level to work out how many queries to make.

We could extend this version of the method more easily by adding or removing levels.

We could further extend this refactoring to remove the methods entirely but I'm leaving that for a later step, wary of being overly distracted.

This gem used to be in Rails Core back in 2007[1] and authored by DDH orginally. It has been forked and looks to have recieved a small amount of steady commits since on it's fork[2] - the one we're using here. They look to be largely supporting new versions of Rails rather.

Before deciding on acts_as_tree, I explored ancestry and acts_as_nested:

1. Ancestry [3] - This gem looks great but doesn't play well at all with using FactoryBot. You have to save a parent without children first or it raises an error. We do a lot of `build` in our seeding and test suites. Our Activity factory for a project, will link to a programme and then a fund. What we'd have to do is a lot of work to always set up the fund before the programme and then before the project. Undoing a lof ot the benefit of FactoryBot - not something I'm comfortable with.
2. Acts_as_tree - This gem looks more complex in it's configuration and feels a heavier tool than we need. That said, I tried it and it failed during the migrations as it isn't expecting UUIDs to be the primary key.

[1] https://github.com/rails/acts_as_tree
[2] https://github.com/amerine/acts_as_tree
[3] https://github.com/stefankroes/ancestry