Move activity show content to tabs
Now that the activity show view has tabs, move the content into the
appropriate tab.

The activity#show action simple redirects to the Financials tab as it
will be used in later work as a 'summary' tab.

During this work it was noted that the `extending_organisation` routes
should be a singular resource so this has been fixed.