Eager load activity parents when querying for programmes, projects or third-party projects
Adding includes provides an optimisation on database queries. On a dummy page I made with a couple of hundred activities at all levels, the time went from:

 (Views: 753.1ms | ActiveRecord: 4.0ms | Allocations: 354833)

To:

(Views: 451.8ms | ActiveRecord: 3.1ms | Allocations: 224901)

Not perfect but an improvement so I think it's worth having.

The Bullet whitelist was added for the unit tests of classes `FindProjectActivities`, `FindProgrammeActivities` and `FindThirdPartyProjectActivities`. All of which require `includes(:parent)` for the feature tasks around them to pass, however they don't themselves exectute those associations. Bullet detects this as an unused eager load, which is true but only within the context in the test. It felt reasonable to whitelist this over exercising the eager load with `result.map(&:parent)` at the end of every test.

I tested that setting the whitelist in this way does not overwrite the global Bullet configuration. I ran 2 tests, the former with the whitelist change, and the latter without. The first file passes and the second fails, confirming that the whitelist is not applied to other tests: `bundle exec rspec --tty spec/services/find_programme_activities_spec.rb spec/services/find_project_activities_spec.rb`