Activity identifiers are only unique between sibling activities
Activities can now have the same identifier, but only when they belong
to different parent activities.

We are loosening up the uniqueness validation on activity identifiers
because we believe it is likely that the project and third-party project
identifiers currently used by delivery partners have the possibility of
being the same.

When using this reference to publish to IATI we think this approach
would continue to work. The fund, programme and project identifiers are
all combined which should guarantee uniqueness across all activities.

NB: We originally added this because we found in practice that programme
identifiers were being reused by BEIS. However, because they would still
share a common fund activity, this change wouldn't actual have fixed
this issue. In that specific case, we chose to change the identifier to
suffix the delivery partner's initials.

Co-authored-by: Tom Hipkin <tomhipkin@gmail.com>