Add basic all activities page
Create the navigation element and page itself.

The action and route were already in-place from a previous feature that
was later removed.

The navigation is only shown to delivery partners, however the route is
not restricted whilst we continue refining these areas in the
application.

Only shows projects and third party projects (level C and D) which meant
the users_can_view_activities_spec needed updating to reflect this.