Redacted activities redact their children
When redacting an activity, all its children activities are redacted as well.

In theory this will only happen for level C activities at the moment, as only
level C & D can be redacted, and only level C can have children. But if we
beging publishing level B to IATI, this will be useful for redacting any children
that level B activities have.