An activity can be asked which level it belongs to
We know what the level of activity is as that is stored directly on the `level` column. What we'd like is a way to find out what the 'level above' this level is, in order to allow users to select that when creating the activity.

Enum values can have their own integer assigned to them in the database, eg. `enum status: { active: 0, archived: 1 }`. Having the integer here would have allowed this method to be _slightly_ improved by removing the need to first cast to a hash of keys. Eg. `Activity.levels[level-1]`. Changing the definition of Enum makes the test suite rather unhappy. As level is such a central concept, and the value it would add to this method is small, I'm building around the existing definition rather than going too deep with this change.