Refactor Find activity services so that the scope is told explicity what user to check
Before this change we were relying on Pundit's hidden use of calling `current_user` to find the user to use in any checks.

This worked since the finders were called in controllers where current_user is available.

I notice that `current_user` was being required by the services as an argument. I think the new reader method took precedent over controller.current_user when called by Pundit. Again this worked but I think this becomes a bit magical and hard to follow.

I ran into this when adding a new helper test that called these services where `current_user` wasn't required for the purposes the test.

`current_user` is a commonly used method so I thought to rename the service parameter to `user` to remove any confusion since we want to be clear that we don't intend to call out to `current_user` at all.