Add a helper method to retreive parent activiti...
..that the user is authorised to see

In order to present a new activity form step that allows the user to connect the new activity to an existing parent we need a way to query for these.

We also want to make sure that only activities the user is authorised to see are the ones that are returned. For this we can reuse the Finder services for the heavy lifting and test that we call the service correctly.

When the activity is a fund there will be no parent activities. Instead of an empty `array` we use Activity.none to return a `Activity::ActiveRecord_Relation`. This ensures that the response of this method will always behave like a query object that can be chained onto, which isn't possible if you return an array.