Add a helper method to help the user select a parent activity
When creating a new activity the new first step will ask them to choose a level for which the activity will be.

Only certain users can create certain levels of activity. I have reused the logic used in Pundit to allow it to continue to be the single control mechanism for what users can and can't do.

If the user is disallowed from creating activity levels in future then this method should automatically reflect that.

The decision to check for both `create?` or `update?` is something we can change. Using either one would work at the moment as they are both equivalant at the moment, this is not always sure to be the case as look at adding different roles to users.