Activity finders can be told not to eager load parent activities
During the refactor of the activity form I was reusing the finder services to allow the user to select from a list of activities. Their `parent` association was not being exercised in this case and Bullet was correctly flagging the efficicency problem.

Leave it to eager loading by default in line with the current behaviour, and allow the user of this class to pass a parameter to prevent this with a boolean.