Add activities index page
This is a simple view of an organisations activities.

The view takes a organisation_id and displays all the activities that
are associated with that organisation through the `organisation`
association.

This means only BIES users will see level A and B activities (Funds and
Programmes) for the moment.

BEIS users can supply a delivery partners organisation id and see those
activities. This functionality is not exposed in the UI as yet, but is
supported by the view, we expect later work to add this UI.

Delivery partners cannot use another organisation id.

The view can be augmented later with additional filters as we learn what
users want to see.

When an unexpected organisation_id is encountered we try to show the
current users own organisation activities instead, those instances might
be:

- no organisation id parameter
- an unknown organisation id
- an empty organisation id
- another organisations id when the user is not a BEIS user

We chose to handle these cases and display the current users activities
as we don't expect to encounter these situations under normal conditions.