Refactor activity specs to use activity factories
We are about to add validation onto activities that requires a means that any activity that isn't a fund, requires a parent.

This means that all places that were using the older, more generic syntax of: `create(:activity, level: xx)`to create something other than a fund produced validation errors.