Activity factory improvement
Funds and programmes should always be created by the BEIS organisation.

Whilst in every test we could provide an organisation:

```
create(:fund_activity, organisation: beis)
create(:programme_activity, organisation: beis)
```

Having to create a BEIS organisation each time can become noisy and is another thing a user of this factory must remember. Given it is always going to be BEIS, we can add it to the factory as a convenience to us.