Remove back links, except on Activity form
The back links are trying to mimic the action of the Back button in the browser
but are increasingly confusing as many pages can be accessed in different ways.

This commit removes them, except for the Back links on the activity create/edit
form as there is no other quick way to get back to the activity details page
without a back link. The back link on the Activity form has been amended to
point at the `details` view of the Activity.