Activities view can be filtered by organisation
A BEIS user can choose a delivery partner organisation and show that
organisations activities.

This is the simplest response to allow BEIS users to view other
organisations activity, we expect this functionality to improve over
time.

As the form that supplies the filter is not tied to a model, the GOVUK
form builder was not used and so the styles are applied manually, see:

https://design-system.service.gov.uk/components/select/