Remove programmes from organisation show page
We are no longer showing programmes in this view, instead they are at
`/activities`.

We still need the programmes table to display on a fund level activities
details tab.

I decided to keep a spec around a delivery partner viewing a programme
if they know the uid, although this feels a bit odd, programmes that the
  delivery partner is the extending organisation of can be viewed even
  though we do not expose this in the interface at the moment.