Child activities are shown in their own tab
Moving child activities to their own tab allows users to view and navigate
the activity hierarchy without having to hunt around the interface.

For this work, what is displayed in the tab remains as before, we may
want to change this in future work.

I've intentionally avoided the language issue by labelling the tab as
generic as possible, we still don't know how or when we will have to
bring the 'level' language into the interface and I feel like 'Child
activities' is actually clearer than 'Level C activities'.