Allow funds with no identifier to be edited
I noticed that a fund activity that is at the identifier step will not
show it's 'add' link, this means a user cannot complete the activity
easily and it may become a dead end.

Looking to fix it I felt we need to do quite a large chunk of work
around this issue and decided to use a simple quick fix to resolve the
main issue by allowing the user to see the edit link, getting the
correct 'add' link seems to need more work.

I also noticed that the activity table and activity presenter were both
trying to localise the level.

I struggled with the `users_can_edit_an_activity` spec as it has
gone thorough an number of changes. I took the time to restructure this
spec to make things a little clearer and closer to reality.

I feel like this spec is starting to get a little deep in contexts as it
is doing a lot, but again, felt this was not time to address the root
issue.