Refactor a create_activity_level_options test for extendability
The test now includes less knowledge of the objects that are returned. This was noticed as the test failed when the object changed to include a new 'description' field. This test is concerned with making sure only the expected options are returned, not with their complete structure - which is left to another test.