Data migration to create a Submission for each delivery partner & fund
This data migration creates a Submission for each DP + fund pairing, for example:

Royal Academy of Engineering has two submissions for GCRF and Newton Fund
UK Space Agency has one submission, for GCRF

There are several delivery partners in the CSV mapping file which are skipped
because they do not have an IATI identifier. This file acts as a record to
indicate which DPs we are currently missing, and to flag that they will need to
be added at a later date.