(fix) Remove tabindex from tabs
We are not using the JS enabled GOVUK frontend tabs so these are
stopping users who navigate with the keyboard from changing tab.

The behaviour in the GOVUK JS tabs is to use the tab key to reach the
active tab and then the arrow keys and return to change the active tab.

I am not confident we can replicate this behaviour quickly or without JS
and feel this solution gives us enough functionality for keyboard users
by letting them use the tab and return key to switch tabs.