Add deadline attribute to Submission, allow Submission editing
Add a Deadline date attribute to Submissions (not required).

BEIS users can edit an existing Submission to add a deadline (and amend the
description field)

Also this commit corrects the location of an error message for the Submission
model. The error shown when an attached Activity is not a fund was in the wrong
location.