Make MigrateUksaGcrfMappings more tolerant of missing data
When running this migration on pentest & staging, we encountered issues where
if a programme is missing, the migration could not run. This is a problem
because we cannot expect each environment to contain exactly the same data.

This fix makes the migration more tolerant of missing programmes & therefore
less likely to break the build on staging.