Changelog entry about the new transparency_identifier
The identifier used when an activity is published to IATI should never be changed
once it has been published. This change introduces a new read-only identifier
field which cannot be edited by users. This is a "first phase" change towards
allowing activities to have an IATI identifier (never changing) for transparency
purposes, and an "internal" identifier for BEIS- and delivery partner-use