Address a flaky test with Submissions
When activating a Submission, a PublicActivity record with the string
`submission.activate` is created. This test is checking for the existence of
this record. However, as part of the test setup, another PublicActivity with the
string `submission.update` is created. Sometimes, when using `.find_by` to
retrieve the PublicActivity records associated with the Submission, the `activate`
record appears first, other times the `update` record appears first and causes
the test to fail.

This changes makes the test consistent by retrieving both PublicActivity records
and testing for the existence of the `submission.activate` records in them, not
just relying on the one returned first.