Rename Submission to Report
Report is the language chosen to for this concept in the service so it
makes sense to keep the model close to the thing it tries to represent.

Also renames the `activate_report` method in the controller as this
method modifies the report the bang makes this clear and as the method
is in the report controller we have 'report scope'.