Add reports index view
This will become the primary place users go to to manage their reports.

For BEIS users both active and inactive reports are shown.

For Delivery partner users only the current active reports are shown.

As we move the report submission journey forward, more report states
will be introduced.

Includes a rename of a spec file to better represent the specs.

Also includes a small change to the default locales, we are striving
for consistently user 'View' rather than 'Show'. I decided to add a view
  key rather than refactor every instance of the show key in the
  codebase.