The Submission association on a Transaction is now non-optional
We have decided after discussion to make the Submission -> Transaction association
required. A Transaction should always belong to a Submission. If no Submission
is available, Transactions cannot be created.

This commit also ensures ingested Transactions get associated to a Submission, and
the ingest will fail if one is not available.