Prevent tests connecting to Redis
In the past we set up `mock_redis` to stand in for Redis in the test
environment. Unfortunately, when we added the JavaScript tests, Redis
gets used once more because the headless browser uses the session store
that is connecting to Redis directly, skipping the mocked version.

Because the Redis session store only accepts a list of servers, rather
than already-instantiated Redis instances, we can't use `mock_redis`
here.

I've replaced `mock_redis` with `fakeredis` gem.  When required in
`rails_helper.rb` it replaces `Redis.new` with a fake version, so this
works everywhere including the session store.