Programme status field added to the activity form
After a user research session, it was identified as a priority for the users
to have the field `programme status` added to the activity form so they can track
the status of each activity. Users preferred this status to the previously used
`IATI status`, as they are more familiar with the terminology of Programme status.
Since IATI status is a requirement for IATI, we will still keep it in the
background to ensure the reports follow the IATI standards. The client provided
us with a mapping chart to convert Programme status into IATI status, which we
will use on a later commit.

This commit introduces the new field in the form, including controllers and views,
validations in the model, codelist and the creation of the column in the database.
It also includes several tests on the specs to assert we achieve the desire
behaviour of this new feature.

Please note that the codes and values for `programme_status` have been added to
the codelist for IATI in `iati.en.yml` despite not belonging to IATI because
we think this improves user experience by allowing the user to see the locale
value for a specific programme status when they see the activity details. The
reason behind it is that there is some logic in the Activity Presenters that use
the Codelist helpers to show the locale values from that file instead of the
original codes.