Placeholder for programme status when ingesting IATI data
After including the new `programme status` field in the activity form,
I found that the tests for ingesting IATI data were failing due to the
validations included in the activity model regarding the required presence
of `programme status`.
We identified that we would need a mapping IATI status-Programme status.
At the moment we only have a mapping for Programme status to IATI status.
As a result, we agreed with the client that for now we will include a
placeholder for programme status on the `ingest_iati_activities` which will be
replaced in the future if a mapping is provided. If not, we will think about
how to handle this data as part of a future card in our backlog on
how to ingest non-IATI data.