Add financial quarter and year to Report
Each report contains the changes made to the data inside RODA for a
specific financial quarter.

In order to show data to users in ways that are meaningful to them, we
store which quarter a report encapsulates. We then use this information
in various ways throughout the interface.

For example, when reviewing the contents of a report we want to see
the actuals and forecast for the report financial quarter in order to
calculate the variance and to help users review the reports contents.

I decided to set the values in the initializer after considering
alternatives like callbacks and service objects. I like the idea that the
values are set regardless of what is going on so we can be sure a Report
has the values when created.

I also set the attributes to read only so they are ignored during any
later update actions.