Show the financial quarter of the report
On the reports tables.

At the moment all reports shown are for the same financial quarter so
this may not be the most helpful way to use this information, but as a
fist step on the path to a complete report - review - submit journey it
gives me confidence that the financial quarters can be shown correctly.