Change the way Report state is stored
Defining the enum as an array has two drawbacks:

1. the values in the DB have no context for developers as they are
   simply numbers
2. less flexibility to allow change as the order of the array matters

By switching to a hash we get around both these drawbacks.

This work also adds an index to state as it is likely the application
will need to sort and display Reports by their state.

We currently have reports in production but they are all 'inactive' so
this work contains a simple data migration to map the stored 0s to
"inactive". Doing this work any later would be far more difficult.