Add the submit journey
Deliver partners can submit their report, confirm and see a completion
page.

The report_state_controller will handle all the state changes a report
goes through.

I chose to use edit and update actions as these felt closer to what is
happening to the report object, we are updating the state.

For the main logic in each action I went with a case block as my
intention is to add the prior state change (inactive -> active) straight
after this work and the rest of the states will follow.

The case block includes a default as an interim for not handling states
that a report might be in - although this is not strictly required, I
wanted to test the Report policy was working in this context and
returning a 401 status.