Users can view submitted reports
Both Delivery partners and BEIS users can view submitted reports.

For Delivery partners, only their own.

For BEIS all organisations.

Moved the inactive reports to the top of the page to reflect the order
in which reports move through different states.

Reduced the size of the table titles as the number of tables on the page
is growing.