Schema migration on `form_state`
After including a new field in the activity form called `programme status` that
replaces the old `status` step in the create activity journey, we realised that
this could be getting errors from activities in which the `form_state` is at the
`status` step.

We first tried to keep the name of the form step as `status` to avoid these errors,
unfortunately `Wicked Wizard` needs the step name to be exactly the same as the
attribute name. Our attribute name needs to be `programme_status` because we are
still keeping the old attribute `status` in the backend for IATI validation
purposes. This left us with the only option of running a schema migration that
will replace the `form_state` to `programme_status` on those activities that
are at a `status` step.