Default `programme status` value for Funds
After including the new field `programme status` on the activity form, the client
asked us to remove this form step only for the creation of Funds/level A activities.
We agreed we will not be asking the users to add a `programme status` when they
create a Fund, but we will set a default value in the background that will help
us deal with validations and legacy data in a better way.

This commit sets the default value of `programme status` to the code 07, which
is `Spend in progress`. This was the value suggested by the client for all new
Funds that get created. In a similar way, RODA will set the `iati status` to the
equivalent for that `programme status` value, according to the mapping we have.