`programme status` form step removed for Funds
Funds (or level A activities) do not require that a user adds a `programme status`.
They are not reported to IATI either, so they do not need to have a `status`.
The client has asked us to remove the `programme status` form step for users
creating Fund level activities, which can help to improve the user experience on
this journey.

This commit adds logic so when a user is creating, editting or viewing a Fund level
activity, they won't need to deal with `programme status`.

On a previous commit we have set a default `programme status` value for Funds,
to avoid potential errors on validations.