Schema migration to backfill `programme_status`
Existing Funds will not have a `programme_status` value, as this field has been
introduced just recently. To avoid potential errors, we have agreed to backfill
this field on existing Funds to the value `Decided` (code `06`).

This commit contains the schema migration that changes the value from `nil` to
`06` for existing Funds.